Size

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 /**
- * main - A program that prints the size of various types on the computer it is compiled and run on
+ * main - A program that prints the size of various types
  * Return: 0
  */
 int main(void)
@@ -10,10 +10,10 @@ int b;
 long int c;
 long long int d;
 float e;
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(a));
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(b));
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(c));
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(d));
-printf("Size of a char: %lu byte(s)\n", (unsigned long) sizeof(e));
+printf("Size of a char: %lu byte(s)\n", sizeof(a));
+printf("Size of a char: %lu byte(s)\n", sizeof(b));
+printf("Size of a char: %lu byte(s)\n", sizeof(c));
+printf("Size of a char: %lu byte(s)\n", sizeof(d));
+printf("Size of a char: %lu byte(s)\n", sizeof(e));
 return (0);
 }
